Quine-McCluskey algorithm
The Karnaugh Map (KM) method of logic simplification works very well for
4
variables or less. It is quick and simple, and can be performed by hand
on paper. It
can be extended to 5 variables or even more, but it becomes far less
intuitive and
more error prone. An alternative technique that is logically equivalent
to Karnaugh
Maps that scales to many variables is known as the Quine-McCluskey (QM)
method.
This technique is tabular and can still be performed by hand on paper.
It is also
relatively simple to automate with a computer. Although other computer
based
techniques have since superseded it (reference), it is still a useful
technique as it is
functioanally equivalent to a Karnaugh Map and can still be performed
manually
on paper.